Be more dynamic in the location of libraries.
First, look for the xtables, iptc, and c libraries in the
standard locations so we don't have hard-coded versions.
Second, because we switched to an extension, the libxtwrapper
code is probably in different places on different architectures
(/usr/lib/python2.7 in i386, /usr/lib64/python2.7 in x86_64).
Therefore, use the python arch sitelib to locate the library.

Signed-off-by: Chris Lalancette <clalance@redhat.com>