Try to load the kernel module.
When iptc is imported try to load the ip_tables kernel module (it
is not loaded automatically).  Raise an exception if this fails
(modprobe returns with anything but 0).  If the module is already
loaded modprobe will just return with 0, so it will do no harm if
we try to load it more than once.