Don't use global constants.
Don't create tables and policies as constants in ip4tc and ip6tc,
since this prevents to import iptc if we don't have root privileges.
This makes it easier to e.g. generate documentation without using
sudo.

This changes the calling semantics somewhat, e.g. to get access to
the filter table use:

    >>> from iptc import Table
    >>> table = Table(Table.FILTER)