Fix Chain._get_rules().
If we create a Rule() inside the loop that traverses iptables rules
in a chain, that will call a method in Table that will mess up the
internal cursor in iptables - thus we will only get back the first
rule in the chain. Fix this by creating Rule objects outside of the
loop.