Implement better alias handling.
When an extension has a 'real_name' attribute, look up that module and
use it as self._module except the 'x6_parse' and 'x6_options' attributes
from the original module (if they are not NULL). The logic in iptables
is that alias modules only use these two for parsing input parameters,
everything else is implemented in the 'real' module.