Try to look up real extension name if init fails.
When we are looking for a match/target and can't seem to find/register
it, try to check for its "real" name (assuming it is an alias). E.g. the
"state" match is now an alias only in the "conntrack" module, and the
init function for registering it is called "libxt_conntrack_init".

This fixes #112.