Set xtables_{matches,targets} to NULL.
When trying to find a match or target, explicitly set `xtables_matches`
and `xtables_targets` to NULL. At the same time, cache matches and
targets we have found.

This way we can use protocol-independent extensions for both IPv4 and
IPv6. When a match or target is not in the cache, libxtables will find
it, and it will be added to the cache.

This fixes #150.